pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "advence-rag"
version = "0.1.0"
description = "Multi-Agent RAG System using Google ADK"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Developer"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "google-adk>=0.3.0",
    "google-genai>=1.0.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0.0",
    "python-magic-bin>=0.4.14",
]

[project.optional-dependencies]
# Vector Store
chroma = ["chromadb>=0.4.0"]

# Document Parsers - 依情境選用
docling = ["docling>=2.0.0"]
unstructured = ["unstructured[all-docs]"]
pymupdf = ["pymupdf4llm>=0.0.10"]
all-parsers = [
    "docling>=2.0.0",
    "unstructured[all-docs]",
    "pymupdf4llm>=0.0.10",
]

# Background Scheduler
scheduler = ["apscheduler>=3.10.0"]

# Reranking
rerank = ["sentence-transformers>=2.2.0"]

# Full installation
full = [
    "chromadb>=0.4.0",
    "pymupdf4llm>=0.0.10",
    "apscheduler>=3.10.0",
    "sentence-transformers>=2.2.0",
]

# Development
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
advence-rag = "advence_rag.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
